requirements.txt
# FastAPI & Web Server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Document parsing
pymupdf>=1.23.0
python-docx>=1.1.0
python-multipart>=0.0.7

# Environment management
python-dotenv>=1.0.1

# Streamlit frontend
streamlit>=1.30.0
requests>=2.31.0

# Groq SDK
groq>=0.22.0

# Pydantic model support
pydantic>=2.11.3
